# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user": {
      "get": {
        "operationId": "UserController_getAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "User limit",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Get users",
            "content": {
              "application/json": {
                "example": {
                  "count": 3,
                  "rows": [
                    {
                      "id": "e5f0f45f-7d23-4c2f-9ff8-a86d9369fdad",
                      "email": "user2@gmail.com",
                      "subscription": "free",
                      "createdAt": "2026-01-05T10:46:01.740Z"
                    },
                    {
                      "id": "c36825e1-1c05-4274-91b1-37489d268709",
                      "email": "user@gmail.com",
                      "subscription": "free",
                      "createdAt": "2026-01-05T10:45:57.113Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "message": "Forbidden resource",
                  "error": "Forbidden",
                  "statusCode": 403
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get one user",
            "content": {
              "application/json": {
                "example": {
                  "id": "e5f0f45f-7d23-4c2f-9ff8-a86d9369fdad",
                  "email": "user2@gmail.com",
                  "subscription": "free",
                  "createdAt": "2026-01-05T10:46:01.740Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "message": "Forbidden resource",
                  "error": "Forbidden",
                  "statusCode": 403
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find one user",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "e5f0f45f-7d23-4c2f-9ff8-a86d9369fdad",
                  "email": "user2@gmail.com",
                  "subscription": "free",
                  "createdAt": "2026-01-05T10:46:01.740Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "message": "Forbidden resource",
                  "error": "Forbidden",
                  "statusCode": 403
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted user successfully",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "message": "The user 879cdf73-993e-40bc-809e-ad8e20cb0bf0 was deleted"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "message": "Forbidden resource",
                  "error": "Forbidden",
                  "statusCode": 403
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fl studio API",
    "description": "Documentation for Fl studio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      },
      "refresh_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Refresh token: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}